+++ src/bsoncxx/lib/bsoncxx/v_noabi/bsoncxx/types/bson_value/value.cpp
@@ -14,7 +14,7 @@
 
 #include <bsoncxx/exception/error_code.hpp>
 #include <bsoncxx/exception/exception.hpp>
-#include <bsoncxx/stdx/make_unique.hpp>
+#include <bsoncxx/private/make_unique.hh>
 #include <bsoncxx/types/bson_value/private/value.hh>
 #include <bsoncxx/types/bson_value/value.hpp>
 #include <bsoncxx/types/private/convert.hh>
@@ -27,19 +27,19 @@ namespace types {
 namespace bson_value {
 
 value::value(b_double v) : value(v.value) {}
-value::value(double v) : _impl{stdx::make_unique<impl>()} {
+value::value(double v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_DOUBLE;
     _impl->_value.value.v_double = v;
 }
 
 value::value(b_int32 v) : value(v.value) {}
-value::value(int32_t v) : _impl{stdx::make_unique<impl>()} {
+value::value(int32_t v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_INT32;
     _impl->_value.value.v_int32 = v;
 }
 
 value::value(b_int64 v) : value(v.value) {}
-value::value(int64_t v) : _impl{stdx::make_unique<impl>()} {
+value::value(int64_t v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_INT64;
     _impl->_value.value.v_int64 = v;
 }
@@ -47,31 +47,31 @@ value::value(int64_t v) : _impl{stdx::make_unique<impl>()} {
 value::value(const char* v) : value(stdx::string_view{v}) {}
 value::value(std::string v) : value(stdx::string_view{v}) {}
 value::value(b_string v) : value(v.value) {}
-value::value(stdx::string_view v) : _impl{stdx::make_unique<impl>()} {
+value::value(stdx::string_view v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_UTF8;
     _impl->_value.value.v_utf8.str = make_copy_for_libbson(v);
     _impl->_value.value.v_utf8.len = static_cast<uint32_t>(v.size());
 }
 
 value::value(b_null) : value(nullptr) {}
-value::value(std::nullptr_t) : _impl{stdx::make_unique<impl>()} {
+value::value(std::nullptr_t) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_NULL;
 }
 
 value::value(b_date v) : value(v.value) {}
-value::value(std::chrono::milliseconds v) : _impl{stdx::make_unique<impl>()} {
+value::value(std::chrono::milliseconds v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_DATE_TIME;
     _impl->_value.value.v_datetime = v.count();
 }
 
 value::value(b_oid v) : value(v.value) {}
-value::value(oid v) : _impl{stdx::make_unique<impl>()} {
+value::value(oid v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_OID;
     std::memcpy(_impl->_value.value.v_oid.bytes, v.bytes(), v.k_oid_length);
 }
 
 value::value(b_bool v) : value(v.value) {}
-value::value(bool v) : _impl{stdx::make_unique<impl>()} {
+value::value(bool v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_BOOL;
     _impl->_value.value.v_bool = v;
 }
@@ -79,7 +79,7 @@ value::value(bool v) : _impl{stdx::make_unique<impl>()} {
 value::value(b_maxkey) : value(type::k_maxkey) {}
 value::value(b_minkey) : value(type::k_minkey) {}
 value::value(b_undefined) : value(type::k_undefined) {}
-value::value(const type id) : _impl{stdx::make_unique<impl>()} {
+value::value(const type id) : _impl{make_unique<impl>()} {
     switch (id) {
         case type::k_minkey:
             _impl->_value.value_type = BSON_TYPE_MINKEY;
@@ -115,8 +115,7 @@ value::value(const type id) : _impl{stdx::make_unique<impl>()} {
 }
 
 value::value(b_regex v) : value(v.regex, v.options) {}
-value::value(stdx::string_view regex, stdx::string_view options)
-    : _impl{stdx::make_unique<impl>()} {
+value::value(stdx::string_view regex, stdx::string_view options) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_REGEX;
     _impl->_value.value.v_regex.regex = make_copy_for_libbson(regex);
     _impl->_value.value.v_regex.options =
@@ -125,7 +124,7 @@ value::value(stdx::string_view regex, stdx::string_view options)
 
 value::value(b_code v) : value(v.type_id, v) {}
 value::value(b_symbol v) : value(v.type_id, v) {}
-value::value(const type id, stdx::string_view v) : _impl{stdx::make_unique<impl>()} {
+value::value(const type id, stdx::string_view v) : _impl{make_unique<impl>()} {
     switch (id) {
         case type::k_regex:
             _impl->_value.value_type = BSON_TYPE_REGEX;
@@ -169,7 +168,7 @@ value::value(const type id, stdx::string_view v) : _impl{stdx::make_unique<impl>
 value::value(b_decimal128 v) : value(v.value) {}
 value::value(decimal128 v) : value(type::k_decimal128, v.high(), v.low()) {}
 value::value(b_timestamp v) : value(v.type_id, v.increment, v.timestamp) {}
-value::value(type id, uint64_t a, uint64_t b) : _impl{stdx::make_unique<impl>()} {
+value::value(type id, uint64_t a, uint64_t b) : _impl{make_unique<impl>()} {
     switch (id) {
         case type::k_decimal128:
             _impl->_value.value_type = BSON_TYPE_DECIMAL128;
@@ -207,7 +206,7 @@ value::value(type id, uint64_t a, uint64_t b) : _impl{stdx::make_unique<impl>()}
 }
 
 value::value(b_dbpointer v) : value(v.collection, v.value) {}
-value::value(stdx::string_view collection, oid value) : _impl{stdx::make_unique<impl>()} {
+value::value(stdx::string_view collection, oid value) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_DBPOINTER;
     _impl->_value.value.v_dbpointer.collection = make_copy_for_libbson(collection);
     _impl->_value.value.v_dbpointer.collection_len = static_cast<uint32_t>(collection.length());
@@ -216,7 +215,7 @@ value::value(stdx::string_view collection, oid value) : _impl{stdx::make_unique<
 
 value::value(b_codewscope v) : value(v.code, v.scope) {}
 value::value(stdx::string_view code, bsoncxx::v_noabi::document::view_or_value scope)
-    : _impl{stdx::make_unique<impl>()} {
+    : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_CODEWSCOPE;
     _impl->_value.value.v_codewscope.code = make_copy_for_libbson(code);
     _impl->_value.value.v_codewscope.code_len = static_cast<uint32_t>(code.length());
@@ -231,7 +230,7 @@ value::value(b_binary v) : value(v.bytes, v.size, v.sub_type) {}
 value::value(std::vector<unsigned char> v, binary_sub_type sub_type)
     : value(v.data(), v.size(), sub_type) {}
 value::value(const uint8_t* data, size_t size, const binary_sub_type sub_type)
-    : _impl{stdx::make_unique<impl>()} {
+    : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_BINARY;
     _impl->_value.value.v_binary.subtype = static_cast<bson_subtype_t>(sub_type);
     _impl->_value.value.v_binary.data_len = static_cast<uint32_t>(size);
@@ -241,7 +240,7 @@ value::value(const uint8_t* data, size_t size, const binary_sub_type sub_type)
 }
 
 value::value(b_document v) : value(v.view()) {}
-value::value(bsoncxx::v_noabi::document::view v) : _impl{stdx::make_unique<impl>()} {
+value::value(bsoncxx::v_noabi::document::view v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_DOCUMENT;
     _impl->_value.value.v_doc.data_len = static_cast<uint32_t>(v.length());
     _impl->_value.value.v_doc.data = static_cast<uint8_t*>(bson_malloc(v.length()));
@@ -249,7 +248,7 @@ value::value(bsoncxx::v_noabi::document::view v) : _impl{stdx::make_unique<impl>
 }
 
 value::value(b_array v) : value(v.value) {}
-value::value(bsoncxx::v_noabi::array::view v) : _impl{stdx::make_unique<impl>()} {
+value::value(bsoncxx::v_noabi::array::view v) : _impl{make_unique<impl>()} {
     _impl->_value.value_type = BSON_TYPE_ARRAY;
     _impl->_value.value.v_doc.data_len = static_cast<uint32_t>(v.length());
     _impl->_value.value.v_doc.data = static_cast<uint8_t*>(bson_malloc(v.length()));
@@ -271,16 +270,16 @@ value::value(const std::uint8_t* raw,
     bson_iter_init_from_data_at_offset(&iter, raw, length, offset, keylen);
     auto value = bson_iter_value(&iter);
 
-    _impl = stdx::make_unique<impl>(value);
+    _impl = make_unique<impl>(value);
 }
 
 value::value(void* internal_value)
-    : _impl(stdx::make_unique<impl>(static_cast<const bson_value_t*>(internal_value))) {}
+    : _impl(make_unique<impl>(static_cast<const bson_value_t*>(internal_value))) {}
 
 value::value(const value& rhs) : value(&rhs._impl->_value) {}
 
 value::value(const bson_value::view& bson_view) {
-    _impl = stdx::make_unique<impl>();
+    _impl = make_unique<impl>();
     convert_to_libbson(&_impl->_value, bson_view);
 }