+++ src/mongocxx/lib/mongocxx/v_noabi/mongocxx/events/command_failed_event.cpp
@@ -15,7 +15,7 @@
 #include <bsoncxx/oid.hpp>
 #include <bsoncxx/private/helpers.hh>
 #include <bsoncxx/private/libbson.hh>
-#include <bsoncxx/stdx/make_unique.hpp>
+#include <bsoncxx/private/make_unique.hh>
 #include <bsoncxx/stdx/optional.hpp>
 
 #include <mongocxx/events/command_failed_event.hpp>