+++ src/mongocxx/lib/mongocxx/v_noabi/mongocxx/read_concern.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include <bsoncxx/builder/basic/document.hpp>
-#include <bsoncxx/stdx/make_unique.hpp>
+#include <bsoncxx/private/make_unique.hh>
 #include <bsoncxx/string/to_string.hpp>
 
 #include <mongocxx/exception/error_code.hpp>
@@ -27,8 +27,7 @@
 namespace mongocxx {
 namespace v_noabi {
 
-read_concern::read_concern()
-    : _impl{bsoncxx::v_noabi::stdx::make_unique<impl>(libmongoc::read_concern_new())} {}
+read_concern::read_concern() : _impl{bsoncxx::make_unique<impl>(libmongoc::read_concern_new())} {}
 
 read_concern::read_concern(std::unique_ptr<impl>&& implementation)
     : _impl{std::move(implementation)} {}
@@ -37,12 +36,11 @@ read_concern::read_concern(read_concern&&) noexcept = default;
 read_concern& read_concern::operator=(read_concern&&) noexcept = default;
 
 read_concern::read_concern(const read_concern& other)
-    : _impl(bsoncxx::v_noabi::stdx::make_unique<impl>(
-          libmongoc::read_concern_copy(other._impl->read_concern_t))) {}
+    : _impl(bsoncxx::make_unique<impl>(libmongoc::read_concern_copy(other._impl->read_concern_t))) {
+}
 
 read_concern& read_concern::operator=(const read_concern& other) {
-    _impl = bsoncxx::v_noabi::stdx::make_unique<impl>(
-        libmongoc::read_concern_copy(other._impl->read_concern_t));
+    _impl = bsoncxx::make_unique<impl>(libmongoc::read_concern_copy(other._impl->read_concern_t));
     return *this;
 }